+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Current
 
 - Improve logging in `wait_and_broadcast_multiple_nodes` for post-mortem analysis
+- `hash(SignedTransactionWithNonce)` now is `SignedTransactionWithNonce.hash`, Ethereum transaction hash 
 
 # 0.24.6